\name{describeData}
\alias{describeData}
\title{Describe metadata}
\usage{
  describeData(data = NULL, file = NULL,
    to.character = FALSE)
}
\arguments{
  \item{data}{}

  \item{file}{}
}
\value{
  None, give metadata about data or file
}
\description{
  .. content for \details{} ..
}
\author{
  ahmadou
}

