\name{getAllMetaData}
\alias{getAllMetaData}
\title{Return some metadata from a gdt file}
\usage{
  getAllMetaData(file)
}
\arguments{
  \item{file}{}
}
\value{
  a list of list which contains all relevant metadata
}
\description{
  Return some metadata from a gdt file
}
\author{
  Ahmadou Dicko <dicko.ahmadou at gmail.com>
}

