\name{getColumnLabel}
\alias{getColumnLabel}
\title{.. content for \description{} (no empty lines) ..}
\usage{
  getColumnLabel(data = NULL, file = NULL, colnames = NULL,
    to.data.frame = FALSE)
}
\arguments{
  \item{data}{}

  \item{file}{}

  \item{colnames}{}

  \item{to.data.frame}{}
}
\value{
  a data.frame or a simple message displaying label and var
  (no object)
}
\description{
  .. content for \details{} ..
}
\author{
  ahmadou
}

