\name{write.gdt.gretldata}
\alias{write.gdt.gretldata}
\title{.. content for \description{} (no empty lines) ..}
\usage{
  write.gdt.gretldata(data, filename, encoding = "UTF-8")
}
\arguments{
  \item{data}{}

  \item{filename}{}

  \item{encoding}{}
}
\value{
  a gdt file
}
\description{
  .. content for \details{} ..
}
\author{
  ahmadou
}

