\name{write.gdt.matrix}
\alias{write.gdt.matrix}
\title{.. content for \description{} (no empty lines) ..}
\usage{
  write.gdt.matrix(data, filename, dataname = "Rdata",
    description = "", varlabel = NULL, obslabels = "false",
    encoding = "UTF-8", startobs = 1, endobs = nrow(data),
    frequency = 1, typeofdata = "cross-section")
}
\arguments{
  \item{data}{}

  \item{filename}{}

  \item{dataname}{}

  \item{description}{}

  \item{varlabel}{}

  \item{obslabels}{}

  \item{encoding}{}

  \item{startobs}{}

  \item{endobs}{}

  \item{frequency}{}

  \item{typeofdata}{}
}
\value{
  a gdt file
}
\description{
  .. content for \details{} ..
}
\author{
  ahmadou
}

