\name{write.gdt.ts}
\alias{write.gdt}
\alias{write.gdt.data.frame}
\alias{write.gdt.matrix}
\alias{write.gdt.ts}
\title{Export R object data to gretl gdt}
\usage{
  write.gdt.ts(data, filename, ...)
}
\arguments{
  \item{data}{}

  \item{filename}{}

  \item{...}{}
}
\value{
  a gdt file
}
\description{
  Export R object data to gretl gdt
}
\author{
  Ahmadou Dicko <dicko.ahmadou at gmail.com>
}

