\name{describeData}
\alias{describeData}
\title{Describe metadata}
\usage{
  describeData(data = NULL, file = NULL,
    to.character = FALSE)
}
\arguments{
  \item{data}{}

  \item{file}{}
}
\value{
  return metadata about gretldata object
}
\description{
  Describe metadata
}
\author{
  Ahmadou Dicko <dicko.ahmadou at gmail.com>
}

