\name{getColumnLabel}
\alias{getColumnLabel}
\title{Get label for variable name}
\usage{
  getColumnLabel(data = NULL, file = NULL, colnames = NULL,
    to.data.frame = FALSE)
}
\arguments{
  \item{data}{}

  \item{file}{}

  \item{colnamesma}{}

  \item{to.data.frame}{}
}
\value{
  a data.frame or a simple message displaying label and var
  (no object)
}
\description{
  Get label for variable name
}
\author{
  Ahmadou Dicko <dicko.ahmadou at gmail.com>
}

