\name{write.gdt.ts}
\alias{write.gdt.ts}
\title{.. content for \description{} (no empty lines) ..}
\usage{
  write.gdt.ts(data, filename, ...)
}
\arguments{
  \item{data}{}

  \item{filename}{}

  \item{...}{}
}
\value{
  a gdt file
}
\description{
  .. content for \details{} ..
}
\author{
  ahmadou
}

